+++ tests/integration/components/registries/partial-registration-modal/component-test.ts
@@ -1,5 +1,6 @@
 import { click, render, settled } from '@ember/test-helpers';
 import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
+import NodeModel from 'ember-osf-web/models/node';
 import { setupRenderingTest } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 import { module, test } from 'qunit';
@@ -9,14 +10,35 @@ module('Integration | Component | partial-registration-modal', hooks => {
     setupMirage(hooks);
 
     test('it renders a hierarchical list with root already selected', async function(assert) {
+        assert.expect(8);
         this.store = this.owner.lookup('service:store');
         const root = server.create('node');
         const child = server.create('node', { parent: root });
         const grandChild = server.create('node', { parent: child });
 
         const rootNode = await this.store.findRecord('node', root.id);
+        const childNode = await this.store.findRecord('node', child.id);
+        const grandChildNode = await this.store.findRecord('node', grandChild.id);
         this.set('rootNode', rootNode);
         this.set('isOpen', false);
+        this.set('onContinue', (selectedNodes: NodeModel[]) => {
+            assert.ok(
+                selectedNodes.length === 3,
+                `expected 3 but instead got ${selectedNodes.length}`,
+            );
+            assert.ok(
+                selectedNodes.includes(rootNode),
+                'root node is included in nodes list sent to onContinue hook',
+            );
+            assert.ok(
+                selectedNodes.includes(childNode),
+                'child node is included in nodes list sent to onContinue hook',
+            );
+            assert.ok(
+                selectedNodes.includes(grandChildNode),
+                'grandchild node is included in nodes list sent to onContinue hook',
+            );
+        });
         await render(hbs`
             <Registries::PartialRegistrationModal::Manager
                 @rootNode={{this.rootNode}}
@@ -24,7 +46,10 @@ module('Integration | Component | partial-registration-modal', hooks => {
                 @renderInPlace={{true}}
                 as |modalManager|
             >
-                <Registries::PartialRegistrationModal @modalManager={{modalManager}} />
+                <Registries::PartialRegistrationModal
+                    @modalManager={{modalManager}}
+                    @onContinue={{action this.onContinue}}
+                />
             </Registries::PartialRegistrationModal::Manager>
         `);
         this.set('isOpen', true);
@@ -33,18 +58,41 @@ module('Integration | Component | partial-registration-modal', hooks => {
         assert.dom(`[data-test-item="${child.id}"]`).exists();
         assert.dom(`[data-test-item="${grandChild.id}"]`).exists();
         assert.dom(`[data-test-item="${root.id}"] input`).isChecked();
+        await click('[data-test-continue-registration-button]');
         this.set('isOpen', false);
         await settled();
     });
 
     test('select child selects parent; deselect parent deselects children', async function(assert) {
+        assert.expect(10);
         this.store = this.owner.lookup('service:store');
         const root = server.create('node');
         const child = server.create('node', { parent: root });
         const grandChild = server.create('node', { parent: child });
+
         const rootNode = await this.store.findRecord('node', root.id);
+        const childNode = await this.store.findRecord('node', child.id);
+        const grandChildNode = await this.store.findRecord('node', grandChild.id);
         this.set('rootNode', rootNode);
         this.set('isOpen', false);
+        this.set('onContinue', (selectedNodes: NodeModel[]) => {
+            assert.ok(
+                selectedNodes.length === 3,
+                `expected 3 but instead got ${selectedNodes.length}`,
+            );
+            assert.ok(
+                selectedNodes.includes(rootNode),
+                'root node is included in nodes list sent to onContinue hook',
+            );
+            assert.ok(
+                selectedNodes.includes(childNode),
+                'child node is included in nodes list sent to onContinue hook',
+            );
+            assert.ok(
+                selectedNodes.includes(grandChildNode),
+                'grandchild node is included in nodes list sent to onContinue hook',
+            );
+        });
         await render(hbs`
             <Registries::PartialRegistrationModal::Manager
                 @rootNode={{this.rootNode}}
@@ -52,7 +100,10 @@ module('Integration | Component | partial-registration-modal', hooks => {
                 @renderInPlace={{true}}
                 as |modalManager|
             >
-                <Registries::PartialRegistrationModal @modalManager={{modalManager}} />
+                <Registries::PartialRegistrationModal
+                    @modalManager={{modalManager}}
+                    @onContinue={{action this.onContinue}}
+                />
             </Registries::PartialRegistrationModal::Manager>
         `);
         this.set('isOpen', true);
@@ -65,18 +116,41 @@ module('Integration | Component | partial-registration-modal', hooks => {
         await click(`[data-test-item="${grandChild.id}"] input`);
         assert.dom(`[data-test-item="${grandChild.id}"] input`).isChecked();
         assert.dom(`[data-test-item="${child.id}"] input`).isChecked();
+        await click('[data-test-continue-registration-button]');
         this.set('isOpen', false);
         await settled();
     });
 
     test('select all works, while clear all should not clear root node', async function(assert) {
+        assert.expect(13);
         this.store = this.owner.lookup('service:store');
         const root = server.create('node');
         const child = server.create('node', { parent: root });
         const grandChild = server.create('node', { parent: child });
+
         const rootNode = await this.store.findRecord('node', root.id);
+        const childNode = await this.store.findRecord('node', child.id);
+        const grandChildNode = await this.store.findRecord('node', grandChild.id);
         this.set('rootNode', rootNode);
         this.set('isOpen', false);
+        this.set('onContinue', (selectedNodes: NodeModel[]) => {
+            assert.ok(
+                selectedNodes.length === 3,
+                `expected 3 but instead got ${selectedNodes.length}`,
+            );
+            assert.ok(
+                selectedNodes.includes(rootNode),
+                'root node is included in nodes list sent to onContinue hook',
+            );
+            assert.ok(
+                selectedNodes.includes(childNode),
+                'child node is included in nodes list sent to onContinue hook',
+            );
+            assert.ok(
+                selectedNodes.includes(grandChildNode),
+                'grandchild node is included in nodes list sent to onContinue hook',
+            );
+        });
         await render(hbs`
             <Registries::PartialRegistrationModal::Manager
                 @rootNode={{this.rootNode}}
@@ -84,7 +158,10 @@ module('Integration | Component | partial-registration-modal', hooks => {
                 @renderInPlace={{true}}
                 as |modalManager|
             >
-                <Registries::PartialRegistrationModal @modalManager={{modalManager}} />
+                <Registries::PartialRegistrationModal
+                    @modalManager={{modalManager}}
+                    @onContinue={{action this.onContinue}}
+                />
             </Registries::PartialRegistrationModal::Manager>
         `);
         this.set('isOpen', true);
@@ -100,6 +177,62 @@ module('Integration | Component | partial-registration-modal', hooks => {
         assert.dom(`[data-test-item="${root.id}"] input`).isChecked();
         assert.dom(`[data-test-item="${grandChild.id}"] input`).isChecked();
         assert.dom(`[data-test-item="${child.id}"] input`).isChecked();
+        await click('[data-test-continue-registration-button]');
+        this.set('isOpen', false);
+        await settled();
+    });
+
+    test('selectedNodes captures deselect children', async function(assert) {
+        assert.expect(8);
+        this.store = this.owner.lookup('service:store');
+        const root = server.create('node');
+        const child = server.create('node', { parent: root });
+        const grandChild = server.create('node', { parent: child });
+
+        const rootNode = await this.store.findRecord('node', root.id);
+        const childNode = await this.store.findRecord('node', child.id);
+        const grandChildNode = await this.store.findRecord('node', grandChild.id);
+        this.set('rootNode', rootNode);
+        this.set('isOpen', false);
+        this.set('onContinue', (selectedNodes: NodeModel[]) => {
+            assert.ok(
+                selectedNodes.length === 1,
+                `expected 1 but instead got ${selectedNodes.length}`,
+            );
+            assert.ok(
+                selectedNodes.includes(rootNode),
+                'root node should be included in nodes list sent to onContinue hook',
+            );
+            assert.notOk(
+                selectedNodes.includes(childNode),
+                'child node should not be included in nodes list sent to onContinue hook',
+            );
+            assert.notOk(
+                selectedNodes.includes(grandChildNode),
+                'grandchild node should not be included in nodes list sent to onContinue hook',
+            );
+        });
+        await render(hbs`
+            <Registries::PartialRegistrationModal::Manager
+                @rootNode={{this.rootNode}}
+                @isOpen={{this.isOpen}}
+                @renderInPlace={{true}}
+                as |modalManager|
+            >
+                <Registries::PartialRegistrationModal
+                    @modalManager={{modalManager}}
+                    @onContinue={{action this.onContinue}}
+                />
+            </Registries::PartialRegistrationModal::Manager>
+        `);
+        this.set('isOpen', true);
+        await settled();
+        assert.dom(`[data-test-item="${grandChild.id}"] input`).isChecked();
+        assert.dom(`[data-test-item="${child.id}"] input`).isChecked();
+        await click(`[data-test-item="${child.id}"] input`);
+        assert.dom(`[data-test-item="${grandChild.id}"] input`).isNotChecked();
+        assert.dom(`[data-test-item="${child.id}"] input`).isNotChecked();
+        await click('[data-test-continue-registration-button]');
         this.set('isOpen', false);
         await settled();
     });