+++ types/ember-cli-chart.d.ts
@@ -1,0 +1,23 @@
+declare module 'ember-cli-chart' {
+    interface DataSet {
+        data?: number[];
+        backgroundColor?: string[];
+    }
+
+    export interface ChartData {
+        labels?: string[];
+        datasets?: DataSet[];
+    }
+
+    export interface ChartOptions {
+        aspectRatio?: number;
+        legend?: {
+            display?: boolean,
+        };
+        onHover?: (_: MouseEvent, shapes: Shape[]) => void;
+    }
+
+    export interface Shape {
+        _index: number;
+    }
+}